Refactor EntityInteractEvent to Refactor PlayerEntityInteractEvent
